api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.4

pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.0.0
